pybackend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
pydantic-settings
python-dotenv
mysql-connector-python
python-multipart